pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sqlens"
version = "0.1.0"
description = "Schema intelligence layer for LLMs — enriched database context, not SQL generation."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Pablo Andrés" }]
keywords = ["sql", "llm", "schema", "context", "rag", "bigquery", "database"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.optional-dependencies]
bigquery = ["google-cloud-bigquery>=3.0"]
postgresql = ["psycopg2-binary>=2.9"]
numpy = ["numpy>=1.24"]
vector = ["chromadb>=0.4", "sentence-transformers>=2.0"]
all = [
    "google-cloud-bigquery>=3.0",
    "numpy>=1.24",
    "chromadb>=0.4",
    "sentence-transformers>=2.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
]

[project.scripts]
sqlens = "sqlens.cli:main"

[project.urls]
Homepage = "https://github.com/pabloandr/sqlens"
Documentation = "https://github.com/pabloandr/sqlens#readme"
Issues = "https://github.com/pabloandr/sqlens/issues"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring real database connections (deselect with '-m \"not integration\"')",
]
